@@ -24,6 +24,7 @@ namespace VimeoDotNet.Authorization
         /// Client Id
         /// </summary>
         private string ClientId { get; }
+
         /// <summary>
         /// Client secret
         /// </summary>
@@ -49,8 +50,10 @@ namespace VimeoDotNet.Authorization
         {
             if (string.IsNullOrWhiteSpace(ClientId))
             {
-                throw new InvalidOperationException("Authorization.ClientId should be a non-null, non-whitespace string");
+                throw new InvalidOperationException(
+                    "Authorization.ClientId should be a non-null, non-whitespace string");
             }
+
             if (string.IsNullOrWhiteSpace(redirectUri))
             {
                 throw new ArgumentException("redirectUri should be a valid Uri");
@@ -71,10 +74,10 @@ namespace VimeoDotNet.Authorization
         {
             try
             {
-              var request = BuildAccessTokenRequest(authorizationCode, redirectUri);
-              var result = await request.ExecuteRequestAsync<AccessTokenResponse>();
-              CheckStatusCodeError(result, "Error getting access token.");
-              return result.Content;
+                var request = BuildAccessTokenRequest(authorizationCode, redirectUri);
+                var result = await request.ExecuteRequestAsync<AccessTokenResponse>();
+                CheckStatusCodeError(result, "Error getting access token.");
+                return result.Content;
             }
             catch (Exception ex)
             {
@@ -82,12 +85,13 @@ namespace VimeoDotNet.Authorization
                 {
                     throw;
                 }
+
                 throw new VimeoApiException("Error getting access token.", ex);
             }
         }
 
         /// <inheritdoc />
-        public Boolean VerifyAccessToken(string accessToken)
+        public bool VerifyAccessToken(string accessToken)
         {
             return VerifyAccessTokenAsync(accessToken).Result;
         }
@@ -97,11 +101,7 @@ namespace VimeoDotNet.Authorization
         {
             var request = GenerateVerifyRequest(accessToken);
             var result = await request.ExecuteRequestAsync();
-            if (result.StatusCode == HttpStatusCode.OK)
-            {
-                return true;
-            }
-            return false;
+            return result.StatusCode == HttpStatusCode.OK;
         }
 
         /// <inheritdoc />
@@ -116,10 +116,11 @@ namespace VimeoDotNet.Authorization
 
         #region Private Methods
 
-        private ApiRequest BuildUnauthenticatedTokenRequest(List<string> scopes = null)
+        private ApiRequest BuildUnauthenticatedTokenRequest(IReadOnlyCollection<string> scopes = null)
         {
             if (string.IsNullOrWhiteSpace(ClientId))
-                throw new InvalidOperationException("Authorization.ClientId should be a non-null, non-whitespace string");
+                throw new InvalidOperationException(
+                    "Authorization.ClientId should be a non-null, non-whitespace string");
             if (string.IsNullOrWhiteSpace(ClientSecret))
                 throw new InvalidOperationException(
                     "Authorization.ClientSecret should be a non-null, non-whitespace string");
@@ -137,6 +138,7 @@ namespace VimeoDotNet.Authorization
             request.Body = new FormUrlEncodedContent(parameters);
             return request;
         }
+
         /// <summary>
         /// Build access token request
         /// </summary>
@@ -149,17 +151,21 @@ namespace VimeoDotNet.Authorization
         {
             if (string.IsNullOrWhiteSpace(ClientId))
             {
-                throw new InvalidOperationException("Authorization.ClientId should be a non-null, non-whitespace string");
+                throw new InvalidOperationException(
+                    "Authorization.ClientId should be a non-null, non-whitespace string");
             }
+
             if (string.IsNullOrWhiteSpace(ClientSecret))
             {
                 throw new InvalidOperationException(
                     "Authorization.ClientSecret should be a non-null, non-whitespace string");
             }
+
             if (string.IsNullOrWhiteSpace(authorizationCode))
             {
                 throw new ArgumentException("authorizationCode should be a non-null, non-whitespace string");
             }
+
             if (string.IsNullOrWhiteSpace(redirectUri))
             {
                 throw new ArgumentException("redirectUri should be a valid Uri");
@@ -180,10 +186,10 @@ namespace VimeoDotNet.Authorization
             return request;
         }
 
-        private IApiRequest GenerateVerifyRequest(string AccessToken)
+        private static IApiRequest GenerateVerifyRequest(string accessToken)
         {
-            IApiRequest request = new ApiRequest(AccessToken);
-            string endpoint = Endpoints.Verify;
+            IApiRequest request = new ApiRequest(accessToken);
+            const string endpoint = Endpoints.Verify;
             request.Method = HttpMethod.Get;
             request.Path = endpoint;
 
@@ -240,9 +246,10 @@ namespace VimeoDotNet.Authorization
                 {
                     sb.Append("&");
                 }
-                
+
                 sb.AppendFormat("{0}={1}", WebUtility.UrlEncode(qsParam.Key), WebUtility.UrlEncode(qsParam.Value));
             }
+
             sb.Insert(0, "?");
             return sb.ToString();
         }
@@ -251,7 +258,7 @@ namespace VimeoDotNet.Authorization
 
         #region Helper Functions
 
-        private void CheckStatusCodeError(IApiResponse response, string message,
+        private static void CheckStatusCodeError(IApiResponse response, string message,
             params HttpStatusCode[] validStatusCodes)
         {
             if (!IsSuccessStatusCode(response.StatusCode) && validStatusCodes != null &&
@@ -262,9 +269,9 @@ namespace VimeoDotNet.Authorization
             }
         }
 
-        private bool IsSuccessStatusCode(HttpStatusCode statusCode)
+        private static bool IsSuccessStatusCode(HttpStatusCode statusCode)
         {
-            var code = (int)statusCode;
+            var code = (int) statusCode;
             return code >= 200 && code < 300;
         }
 
