@@ -1,16 +1,9 @@
 ﻿namespace VimeoDotNet.Authorization
 {
-    /// <summary>
-    /// Authorization client factory
-    /// </summary>
+    /// <inheritdoc />
     public class AuthorizationClientFactory : IAuthorizationClientFactory
     {
-        /// <summary>
-        /// Return authorization client
-        /// </summary>
-        /// <param name="clientId">ClientId</param>
-        /// <param name="clientSecret">ClientSecret</param>
-        /// <returns>Authorization client</returns>
+        /// <inheritdoc />
         public IAuthorizationClient GetAuthorizationClient(string clientId, string clientSecret)
         {
             return new AuthorizationClient(clientId, clientSecret);
