@@ -1,10 +1,17 @@
-﻿namespace VimeoDotNet.Constants
+﻿// ReSharper disable InconsistentNaming
+using JetBrains.Annotations;
+
+namespace VimeoDotNet.Constants
 {
     internal static class ApiVersions
     {
+        [PublicAPI]
         public const string v2_0 = "version=2.0";
+        [PublicAPI]
         public const string v2_5 = "version=2.5";
+        [PublicAPI]
         public const string v3_0 = "version=3.0";
+        [PublicAPI]
         public const string v3_2 = "version=3.2";
     }
 }
