@@ -6,17 +6,15 @@ namespace VimeoDotNet.Enums
     public enum VideoCommentsEnum
     {
         /// <summary>
-        /// Unknown
-        /// </summary>
-        Unknown,
-        /// <summary>
         /// Anybody
         /// </summary>
         Anybody,
+
         /// <summary>
         /// Nobody
         /// </summary>
         Nobody,
+
         /// <summary>
         /// Only contacts
         /// </summary>
