@@ -6,17 +6,15 @@
     public enum VideoEmbedPrivacyEnum
     {
         /// <summary>
-        /// Unknown
-        /// </summary>
-        Unknown,
-        /// <summary>
         /// Public
         /// </summary>
         Public,
+
         /// <summary>
         /// Private
         /// </summary>
         Private,
+
         /// <summary>
         /// Whitelist domain
         /// </summary>
