@@ -6,33 +6,35 @@
     public enum VideoPrivacyEnum
     {
         /// <summary>
-        /// Unknown
-        /// </summary>
-        Unknown,
-        /// <summary>
         /// Nobody
         /// </summary>
         Nobody,
+
         /// <summary>
         /// Anybody
         /// </summary>
         Anybody,
+
         /// <summary>
         /// Only contacts
         /// </summary>
         Contacts,
+
         /// <summary>
         /// Only users
         /// </summary>
         Users,
+
         /// <summary>
         /// Password
         /// </summary>
         Password,
+
         /// <summary>
         /// Disable
         /// </summary>
         Disable,
+
         /// <summary>
         /// Unlisted
         /// </summary>
