@@ -1,49 +1,49 @@
 ﻿using System;
-using System.Runtime.Serialization;
+using JetBrains.Annotations;
 using VimeoDotNet.Net;
 
 namespace VimeoDotNet.Exceptions
 {
-    [Serializable]
     internal class VimeoUploadException : VimeoApiException
     {
-        [NonSerialized] private IUploadRequest _request;
-
+        [PublicAPI]
         public VimeoUploadException()
         {
         }
 
+        [PublicAPI]
         public VimeoUploadException(IUploadRequest request)
         {
             Request = request;
         }
 
+        [PublicAPI]
         public VimeoUploadException(string message)
             : base(message)
         {
         }
 
+        [PublicAPI]
         public VimeoUploadException(string message, IUploadRequest request)
             : base(message)
         {
             Request = request;
         }
 
+        [PublicAPI]
         public VimeoUploadException(string message, Exception innerException)
             : base(message, innerException)
         {
         }
 
+        [PublicAPI]
         public VimeoUploadException(string message, IUploadRequest request, Exception innerException)
             : base(message, innerException)
         {
             Request = request;
         }
 
-        public IUploadRequest Request
-        {
-            get { return _request; }
-            set { _request = value; }
-        }
+        [PublicAPI]
+        public IUploadRequest Request { get; set; }
     }
 }
