@@ -1,87 +1,113 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Text.RegularExpressions;
+﻿using System.Text.RegularExpressions;
+using JetBrains.Annotations;
+using Newtonsoft.Json;
 
 namespace VimeoDotNet.Models
 {
     /// <summary>
     /// Album model
     /// </summary>
-    [Serializable]
     public class Album
     {
         /// <summary>
         /// URI
         /// </summary>
-        public string uri { get; set; }
+        [PublicAPI]
+        [JsonProperty(PropertyName = "uri")]
+        public string Uri { get; set; }
 
         /// <summary>
         /// User
         /// </summary>
-        public User user { get; set; }
+
+        [PublicAPI]
+        [JsonProperty(PropertyName = "user")]
+        public User User { get; set; }
 
         /// <summary>
         /// Name
         /// </summary>
-        public string name { get; set; }
+        [PublicAPI]
+        [JsonProperty(PropertyName = "name")]
+        public string Name { get; set; }
 
         /// <summary>
         /// Description
         /// </summary>
-        public string description { get; set; }
+        [PublicAPI]
+        [JsonProperty(PropertyName = "description")]
+        public string Description { get; set; }
 
         /// <summary>
         /// Link
         /// </summary>
-        public string link { get; set; }
+        [PublicAPI]
+        [JsonProperty(PropertyName = "link")]
+        public string Link { get; set; }
 
         /// <summary>
         /// Duration
         /// </summary>
-        public int duration { get; set; }
+        [PublicAPI]
+        [JsonProperty(PropertyName = "duration")]
+        public int Duration { get; set; }
 
         /// <summary>
         /// CreatedTime
         /// </summary>
-        public string created_time { get; set; }
+        [PublicAPI]
+        [JsonProperty(PropertyName = "created_time")]
+        public string CreatedTime { get; set; }
 
         /// <summary>
         /// Pictures
         /// </summary>
-        public Pictures pictures { get; set; }
+        [PublicAPI]
+        [JsonProperty(PropertyName = "pictures")]
+        public Pictures Pictures { get; set; }
 
         /// <summary>
         /// Privacy
         /// </summary>
-        public Privacy privacy { get; set; }
+        [PublicAPI]
+        [JsonProperty(PropertyName = "privacy")]
+        public Privacy Privacy { get; set; }
 
         /// <summary>
         /// Stats
         /// </summary>
-        public AlbumStats stats { get; set; }
+        [PublicAPI]
+        [JsonProperty(PropertyName = "stats")]
+        public AlbumStats Stats { get; set; }
 
         /// <summary>
         /// Metadata
         /// </summary>
-        public AlbumMetadata metadata { get; set; }
+        [PublicAPI]
+        [JsonProperty(PropertyName = "metadata")]
+        public AlbumMetadata Metadata { get; set; }
 
-		/// <summary>
-		/// Return album id if exists
-		/// </summary>
-		/// <returns>AlbumId or null</returns>
-		public long? GetAlbumId()
-		{
-			if (String.IsNullOrEmpty(uri)) { return null; }
+        /// <summary>
+        /// Return album id if exists
+        /// </summary>
+        /// <returns>AlbumId or null</returns>
+        [PublicAPI]
+        public long? GetAlbumId()
+        {
+            if (string.IsNullOrEmpty(Uri))
+            {
+                return null;
+            }
 
-			var match = regexAlbumUri.Match(uri);
-			if (match.Success)
-			{
-				return long.Parse(match.Groups["albumId"].Value);
-			}
+            var match = RegexAlbumUri.Match(Uri);
+            if (match.Success)
+            {
+                return long.Parse(match.Groups["albumId"].Value);
+            }
 
-			return null;
-		}
+            return null;
+        }
 
-		private static readonly Regex regexAlbumUri = new Regex(@"/albums/(?<albumId>\d+)/?$");
+        private static readonly Regex RegexAlbumUri = new Regex(@"/albums/(?<albumId>\d+)/?$");
     }
 }
