@@ -1,5 +1,7 @@
 ﻿using System;
 using System.Collections.Generic;
+using JetBrains.Annotations;
+using Newtonsoft.Json;
 using VimeoDotNet.Enums;
 using VimeoDotNet.Helpers;
 
@@ -8,7 +10,6 @@ namespace VimeoDotNet.Models
     /// <summary>
     /// File model
     /// </summary>
-    [Serializable]
     public class File
     {
         private static readonly IDictionary<string, string> QualityMappings = new Dictionary<string, string>
@@ -22,50 +23,67 @@ namespace VimeoDotNet.Models
         /// <summary>
         /// Quality
         /// </summary>
-        public string quality { get; set; }
+        [PublicAPI]
+        [JsonProperty(PropertyName = "quality")]
+        public string Quality { get; set; }
 
         /// <summary>
         /// Type
         /// </summary>
-        public string type { get; set; }
+        [PublicAPI]
+        [JsonProperty(PropertyName = "type")]
+        public string Type { get; set; }
 
         /// <summary>
         /// Width
         /// </summary>
-        public int width { get; set; }
+        [PublicAPI]
+        [JsonProperty(PropertyName = "width")]
+        public int Width { get; set; }
 
         /// <summary>
         /// Height
         /// </summary>
-        public int height { get; set; }
+        [PublicAPI]
+        [JsonProperty(PropertyName = "height")]
+        public int Height { get; set; }
 
         /// <summary>
         /// File size
         /// </summary>
-        public long size { get; set; }
+        [PublicAPI]
+        [JsonProperty(PropertyName = "size")]
+        public long Size { get; set; }
 
         /// <summary>
         /// Expires
         /// </summary>
-        public DateTime? expires { get; set; }
+        [PublicAPI]
+        [JsonProperty(PropertyName = "expires")]
+        public DateTime? Expires { get; set; }
 
         /// <summary>
         /// Link
         /// </summary>
-        public string link { get; set; }
+        [PublicAPI]
+        [JsonProperty(PropertyName = "link")]
+        public string Link { get; set; }
 
         /// <summary>
         /// LinkSecure
         /// </summary>
-        public string link_secure { get; set; }
+        [PublicAPI]
+        [JsonProperty(PropertyName = "link_secure")]
+        public string LinkSecure { get; set; }
 
         /// <summary>
         /// FileQuality
         /// </summary>
+        [PublicAPI]
         public FileQualityEnum FileQuality
         {
-            get { return ModelHelpers.GetEnumValue<FileQualityEnum>(quality, QualityMappings); }
-            set { quality = ModelHelpers.GetEnumString(value, QualityMappings); }
+            get => ModelHelpers.GetEnumValue<FileQualityEnum>(Quality, QualityMappings);
+            set => Quality = ModelHelpers.GetEnumString(value, QualityMappings);
         }
     }
 }
