@@ -1,26 +1,33 @@
 ﻿using System;
+using JetBrains.Annotations;
+using Newtonsoft.Json;
 
 namespace VimeoDotNet.Models
 {
     /// <summary>
     /// Follow
     /// </summary>
-    [Serializable]
     public class Follow
     {
         /// <summary>
         /// Added
         /// </summary>
-        public bool added { get; set; }
+        [PublicAPI]
+        [JsonProperty(PropertyName = "added")]
+        public bool Added { get; set; }
 
         /// <summary>
         /// Added time
         /// </summary>
-        public DateTime? added_time { get; set; }
+        [PublicAPI]
+        [JsonProperty(PropertyName = "added_time")]
+        public DateTime? AddedTime { get; set; }
 
         /// <summary>
         /// URI
         /// </summary>
-        public string uri { get; set; }
+        [PublicAPI]
+        [JsonProperty(PropertyName = "uri")]
+        public string Uri { get; set; }
     }
 }
