@@ -1,24 +1,32 @@
-﻿using System;
+﻿using JetBrains.Annotations;
+using Newtonsoft.Json;
 
 namespace VimeoDotNet.Models
 {
     /// <summary>
     /// Preset logos
     /// </summary>
-    [Serializable]
     public class PresetLogos
     {
         /// <summary>
         /// Vimeo
         /// </summary>
-        public bool vimeo { get; set; }
+        [PublicAPI]
+        [JsonProperty(PropertyName = "vimeo")]
+        public bool Vimeo { get; set; }
+
         /// <summary>
         /// Custom
         /// </summary>
-        public bool custom { get; set; }
+        [PublicAPI]
+        [JsonProperty(PropertyName = "custom")]
+        public bool Custom { get; set; }
+
         /// <summary>
         /// Sticky custom
         /// </summary>
-        public bool sticky_custom { get; set; }
+        [PublicAPI]
+        [JsonProperty(PropertyName = "sticky_custom")]
+        public bool StickyCustom { get; set; }
     }
 }
