@@ -1,24 +1,32 @@
-﻿using System;
+﻿using JetBrains.Annotations;
+using Newtonsoft.Json;
 
 namespace VimeoDotNet.Models
 {
     /// <summary>
     /// User upload quota
     /// </summary>
-    [Serializable]
     public class UserUploadQuota
     {
         /// <summary>
         /// Space
         /// </summary>
-        public Space space { get; set; }
+        [PublicAPI]
+        [JsonProperty(PropertyName = "space")]
+        public Space Space { get; set; }
+
         /// <summary>
         /// Resets
         /// </summary>
-        public int resets { get; set; }
+        [PublicAPI]
+        [JsonProperty(PropertyName = "resets")]
+        public int Resets { get; set; }
+
         /// <summary>
         /// Quota
         /// </summary>
-        public UserQuota quota { get; set; }
+        [PublicAPI]
+        [JsonProperty(PropertyName = "quota")]
+        public UserQuota Quota { get; set; }
     }
 }
