@@ -1,52 +1,46 @@
-﻿using System;
-using System.Collections.Generic;
+﻿using JetBrains.Annotations;
+using Newtonsoft.Json;
 
 namespace VimeoDotNet.Models
 {
     /// <summary>
     /// Video connections
     /// </summary>
-    [Serializable]
     public class VideoConnections
     {
         /// <summary>
         /// Comments
         /// </summary>
-        public VideoConnectionsEntry comments { get; set; }
+        [PublicAPI]
+        [JsonProperty(PropertyName = "comments")]
+        public VideoConnectionsEntry Comments { get; set; }
+
         /// <summary>
         /// Credits
         /// </summary>
-        public VideoConnectionsEntry credits { get; set; }
+        [PublicAPI]
+        [JsonProperty(PropertyName = "credits")]
+        public VideoConnectionsEntry Credits { get; set; }
+
         /// <summary>
         /// Likes
         /// </summary>
-        public VideoConnectionsEntry likes { get; set; }
+        [PublicAPI]
+        [JsonProperty(PropertyName = "likes")]
+        public VideoConnectionsEntry Likes { get; set; }
+
         /// <summary>
         /// Presets
         /// </summary>
-        public VideoConnectionsEntry puctires { get; set; }
+        [PublicAPI]
+        [JsonProperty(PropertyName = "puctires")]
+        public VideoConnectionsEntry Puctires { get; set; }
+
         /// <summary>
         /// Text tracks
         /// </summary>
-        public TextTracks texttracks { get; set; }
-    }
-
-    /// <summary>
-    /// Video connections entry
-    /// </summary>
-    public class VideoConnectionsEntry
-    {
-        /// <summary>
-        /// URI
-        /// </summary>
-        public string uri { get; set; }
-        /// <summary>
-        /// Options
-        /// </summary>
-        public List<string> options { get; set; }
-        /// <summary>
-        /// Total
-        /// </summary>
-        public int total { get; set; }
+        [PublicAPI]
+        [JsonProperty(PropertyName = "texttracks")]
+        public TextTracks Texttracks { get; set; }
     }
 }
