@@ -1,106 +1,127 @@
-﻿using System;
-using System.Collections.Generic;
+﻿using System.Collections.Generic;
+using JetBrains.Annotations;
+using Newtonsoft.Json;
+using Newtonsoft.Json.Converters;
 using VimeoDotNet.Enums;
 
 namespace VimeoDotNet.Parameters
 {
-	/// <summary>
-	/// GetAlbumsSortOption
-	/// </summary>
-	public enum GetAlbumsSortOption
-	{
-		/// <summary>
-		/// Date
-		/// </summary>
-		Date,
-		/// <summary>
-		/// Alphabetical
-		/// </summary>
-		Alphabetical,
-		/// <summary>
-		/// Videos
-		/// </summary>
-		Videos,
-		/// <summary>
-		/// Duration
-		/// </summary>
-		Duration
-	}
+    /// <summary>
+    /// GetAlbumsSortOption
+    /// </summary>
+    [PublicAPI]
+    public enum GetAlbumsSortOption
+    {
+        /// <summary>
+        /// Date
+        /// </summary>
+        Date,
 
-	/// <summary>
-	/// Get albums sort direction option
-	/// </summary>
-	public enum GetAlbumsSortDirectionOption
-	{
-		/// <summary>
-		/// Ascending
-		/// </summary>
-		Asc,
-		/// <summary>
-		/// Descending
-		/// </summary>
-		Desc
-	}
+        /// <summary>
+        /// Alphabetical
+        /// </summary>
+        Alphabetical,
 
-	/// <summary>
-	/// Get albums parameters
-	/// </summary>
-	public class GetAlbumsParameters : IParameterProvider
-	{
-		/// <summary>
-		/// Page
-		/// </summary>
-		public int? Page { get; set; }
+        /// <summary>
+        /// Videos
+        /// </summary>
+        Videos,
 
-		/// <summary>
-		/// Per page
-		/// </summary>
-		public int? PerPage { get; set; }
+        /// <summary>
+        /// Duration
+        /// </summary>
+        Duration
+    }
 
-		/// <summary>
-		/// Query
-		/// </summary>
-		public string Query { get; set; }
+    /// <summary>
+    /// Get albums sort direction option
+    /// </summary>
+    [PublicAPI]
+    public enum GetAlbumsSortDirectionOption
+    {
+        /// <summary>
+        /// Ascending
+        /// </summary>
+        Asc,
 
-		/// <summary>
-		/// Sort
-		/// </summary>
-		public GetAlbumsSortOption? Sort { get; set; }
+        /// <summary>
+        /// Descending
+        /// </summary>
+        Desc
+    }
 
-		/// <summary>
-		/// Direction
-		/// </summary>
-		public GetAlbumsSortDirectionOption? Direction { get; set; }
+    /// <inheritdoc />
+    public class GetAlbumsParameters : IParameterProvider
+    {
+        /// <summary>
+        /// Page
+        /// </summary>
+        [PublicAPI]
+        public int? Page { get; set; }
 
-	    /// <summary>
-	    /// Performs validation and returns a description of the first error encountered.
-	    /// </summary>
-	    /// <returns>Description of first error, or null if none found.</returns>
-	    public string ValidationError()
-		{
-			if (PerPage > 50)
-				return "Maximum number of items allowed per page is 50.";
+        /// <summary>
+        /// Per page
+        /// </summary>
+        [PublicAPI]
+        public int? PerPage { get; set; }
 
-			return null;
-		}
+        /// <summary>
+        /// Query
+        /// </summary>
+        [PublicAPI]
+        public string Query { get; set; }
 
-	    /// <summary>
-	    /// Provides universal interface to retrieve parameter values.
-	    /// </summary>
-	    /// <returns>Returns all parameters as name/value pairs.</returns>
-	    public IDictionary<string, string> GetParameterValues()
-		{
-			Dictionary<string, string> parameterValues = new Dictionary<string, string>();
+        /// <summary>
+        /// Sort
+        /// </summary>
+        [PublicAPI]
+        [JsonConverter(typeof(StringEnumConverter))]
+        public GetAlbumsSortOption? Sort { get; set; }
 
-			if (Page.HasValue) { parameterValues.Add("page", Page.Value.ToString()); }
-			if (PerPage.HasValue) { parameterValues.Add("per_page", PerPage.Value.ToString()); }
-			if (!String.IsNullOrEmpty(Query)) { parameterValues.Add("query", Query); }
-			if (Sort.HasValue) { parameterValues.Add("sort", Sort.Value.GetParameterValue()); }
-			if (Direction.HasValue) { parameterValues.Add("direction", Direction.Value.GetParameterValue()); }
+        /// <summary>
+        /// Direction
+        /// </summary>
+        [PublicAPI]
+        [JsonConverter(typeof(StringEnumConverter))]
+        public GetAlbumsSortDirectionOption? Direction { get; set; }
 
-			if (parameterValues.Keys.Count > 0) { return parameterValues; }
+        /// <inheritdoc />
+        public string ValidationError()
+        {
+            return PerPage > 50 ? "Maximum number of items allowed per page is 50." : null;
+        }
 
-			return null;
-		}
-	}
+        /// <inheritdoc />
+        public IDictionary<string, string> GetParameterValues()
+        {
+            var parameterValues = new Dictionary<string, string>();
+
+            if (Page.HasValue)
+            {
+                parameterValues.Add("page", Page.Value.ToString());
+            }
+
+            if (PerPage.HasValue)
+            {
+                parameterValues.Add("per_page", PerPage.Value.ToString());
+            }
+
+            if (!string.IsNullOrEmpty(Query))
+            {
+                parameterValues.Add("query", Query);
+            }
+
+            if (Sort.HasValue)
+            {
+                parameterValues.Add("sort", Sort.Value.GetParameterValue());
+            }
+
+            if (Direction.HasValue)
+            {
+                parameterValues.Add("direction", Direction.Value.GetParameterValue());
+            }
+
+            return parameterValues.Keys.Count > 0 ? parameterValues : null;
+        }
+    }
 }
