@@ -13,12 +13,7 @@ namespace VimeoDotNet
 {
     public partial class VimeoClient
     {
-        /// <summary>
-        /// Get text tracks asynchronously
-        /// </summary>
-        /// <param name="videoId">VideoId</param>
-        /// <returns>Return text tracks</returns>
-        ///
+        /// <inheritdoc />
         public async Task<TextTracks> GetTextTracksAsync(long videoId)
         {
             try
@@ -28,11 +23,7 @@ namespace VimeoDotNet
                 UpdateRateLimit(response);
                 CheckStatusCodeError(response, "Error retrieving text tracks for video.", HttpStatusCode.NotFound);
 
-                if (response.StatusCode == HttpStatusCode.NotFound)
-                {
-                    return null;
-                }
-                return response.Content;
+                return response.StatusCode == HttpStatusCode.NotFound ? null : response.Content;
             }
             catch (Exception ex)
             {
@@ -40,16 +31,12 @@ namespace VimeoDotNet
                 {
                     throw;
                 }
+
                 throw new VimeoApiException("Error retrieving text tracks for video.", ex);
             }
         }
 
-        /// <summary>
-        /// Get text track asynchronously
-        /// </summary>
-        /// <param name="videoId">VideoId</param>
-        /// <param name="trackId">TrackId</param>
-        /// <returns>Return text track</returns>
+        /// <inheritdoc />
         public async Task<TextTrack> GetTextTrackAsync(long videoId, long trackId)
         {
             try
@@ -59,11 +46,7 @@ namespace VimeoDotNet
                 UpdateRateLimit(response);
                 CheckStatusCodeError(response, "Error retrieving text track for video.", HttpStatusCode.NotFound);
 
-                if (response.StatusCode == HttpStatusCode.NotFound)
-                {
-                    return null;
-                }
-                return response.Content;
+                return response.StatusCode == HttpStatusCode.NotFound ? null : response.Content;
             }
             catch (Exception ex)
             {
@@ -71,34 +54,30 @@ namespace VimeoDotNet
                 {
                     throw;
                 }
+
                 throw new VimeoApiException("Error retrieving text track for video.", ex);
             }
         }
 
-        /// <summary>
-        /// Upload new text track file asynchronously
-        /// </summary>
-        /// <param name="fileContent">File content</param>
-        /// <param name="videoId">VideoId</param>
-        /// <param name="track">Track</param>
-        /// <returns>New text track</returns>
+        /// <inheritdoc />
         public async Task<TextTrack> UploadTextTrackFileAsync(IBinaryContent fileContent, long videoId, TextTrack track)
         {
             if (!fileContent.Data.CanRead)
             {
                 throw new ArgumentException("fileContent should be readable");
             }
+
             if (fileContent.Data.CanSeek && fileContent.Data.Position > 0)
             {
                 fileContent.Data.Position = 0;
             }
 
-            TextTrack ticket = await GetUploadTextTrackTicketAsync(videoId, track);
-            IApiRequest request = ApiRequestFactory.GetApiRequest();
+            var ticket = await GetUploadTextTrackTicketAsync(videoId, track);
+            var request = _apiRequestFactory.GetApiRequest();
             request.Method = HttpMethod.Put;
             request.ExcludeAuthorizationHeader = true;
-            request.Path = ticket.link;
-            
+            request.Path = ticket.Link;
+
             request.Body = new ByteArrayContent(await fileContent.ReadAllAsync());
 
             var response = await request.ExecuteRequestAsync();
@@ -107,13 +86,7 @@ namespace VimeoDotNet
             return ticket;
         }
 
-        /// <summary>
-        /// Update text track asynchronously
-        /// </summary>
-        /// <param name="videoId">VideoId</param>
-        /// <param name="trackId">TrackId</param>
-        /// <param name="track">TextTrack</param>
-        /// <returns>Updated text track</returns>
+        /// <inheritdoc />
         public async Task<TextTrack> UpdateTextTrackAsync(long videoId, long trackId, TextTrack track)
         {
             try
@@ -123,11 +96,7 @@ namespace VimeoDotNet
                 UpdateRateLimit(response);
                 CheckStatusCodeError(response, "Error updating text track for video.", HttpStatusCode.NotFound);
 
-                if (response.StatusCode == HttpStatusCode.NotFound)
-                {
-                    return null;
-                }
-                return response.Content;
+                return response.StatusCode == HttpStatusCode.NotFound ? null : response.Content;
             }
             catch (Exception ex)
             {
@@ -135,16 +104,12 @@ namespace VimeoDotNet
                 {
                     throw;
                 }
+
                 throw new VimeoApiException("Error updating text track for video.", ex);
             }
         }
 
-        /// <summary>
-        /// Delete text track asynchronously
-        /// </summary>
-        /// <param name="videoId">VideoId</param>
-        /// <param name="trackId">TrackId</param>
-        /// <returns></returns>
+        /// <inheritdoc />
         public async Task DeleteTextTrackAsync(long videoId, long trackId)
         {
             try
@@ -160,6 +125,7 @@ namespace VimeoDotNet
                 {
                     throw;
                 }
+
                 throw new VimeoApiException("Error updating text track for video.", ex);
             }
         }
@@ -181,44 +147,35 @@ namespace VimeoDotNet
                 {
                     throw;
                 }
+
                 throw new VimeoUploadException("Error generating upload text track ticket.", null, ex);
             }
         }
 
         private IApiRequest GenerateUploadTextTrackTicketRequest(long clipId, TextTrack track)
         {
-            ThrowIfUnauthorized();
-
-            IApiRequest request = ApiRequestFactory.GetApiRequest(AccessToken);
+            var request = _apiRequestFactory.GetApiRequest(AccessToken);
             request.Method = HttpMethod.Post;
             request.Path = Endpoints.TextTracks;
             request.UrlSegments.Add("clipId", clipId.ToString());
-
-            if (track != null)
+            if (track == null)
+                return request;
+            var parameters = new Dictionary<string, string>
             {
-                request.Query.Add("active", track.active.ToString().ToLower());
-                if (track.name != null)
-                {
-                    request.Query.Add("name", track.name);
-                }
-                if (track.language != null)
-                {
-                    request.Query.Add("language", track.language);
-                }
-                if (track.type != null)
-                {
-                    request.Query.Add("type", track.type);
-                }
-            }
-
+                ["active"] = track.Active.ToString().ToLower(),
+                ["name"] = track.Name,
+                ["language"] = track.Language,
+                ["type"] = track.Type.ToString().ToLowerInvariant()
+            };
+            request.Body = new FormUrlEncodedContent(parameters);
             return request;
         }
 
-        private IApiRequest GenerateUpdateTextTrackRequest(long clipId, long trackId, [NotNull]TextTrack track)
+        private IApiRequest GenerateUpdateTextTrackRequest(long clipId, long trackId, [NotNull] TextTrack track)
         {
             ThrowIfUnauthorized();
 
-            var request = ApiRequestFactory.GetApiRequest(AccessToken);
+            var request = _apiRequestFactory.GetApiRequest(AccessToken);
             request.Method = new HttpMethod("PATCH");
             request.Path = Endpoints.TextTrack;
             request.UrlSegments.Add("clipId", clipId.ToString());
@@ -226,20 +183,19 @@ namespace VimeoDotNet
 
             var parameters = new Dictionary<string, string>
             {
-                ["active"] = track.active.ToString().ToLower()
+                ["active"] = track.Active.ToString().ToLower()
             };
-            if (track.name != null)
+            if (track.Name != null)
             {
-                parameters["name"] = track.name;
+                parameters["name"] = track.Name;
             }
-            if (track.language != null)
+
+            if (track.Language != null)
             {
-                parameters["language"] =track.language;
+                parameters["language"] = track.Language;
             }
-            if (track.type != null)
-            {
-                parameters["type"] = track.type;
-            }
+
+            parameters["type"] = track.Type.ToString().ToLowerInvariant();
             request.Body = new FormUrlEncodedContent(parameters);
 
             return request;
@@ -249,8 +205,8 @@ namespace VimeoDotNet
         {
             ThrowIfUnauthorized();
 
-            IApiRequest request = ApiRequestFactory.GetApiRequest(AccessToken);
-            string endpoint = trackId.HasValue ? Endpoints.TextTrack : Endpoints.TextTracks;
+            var request = _apiRequestFactory.GetApiRequest(AccessToken);
+            var endpoint = trackId.HasValue ? Endpoints.TextTrack : Endpoints.TextTracks;
             request.Method = HttpMethod.Get;
             request.Path = endpoint;
 
@@ -259,6 +215,7 @@ namespace VimeoDotNet
             {
                 request.UrlSegments.Add("trackId", trackId.ToString());
             }
+
             return request;
         }
 
@@ -266,8 +223,8 @@ namespace VimeoDotNet
         {
             ThrowIfUnauthorized();
 
-            IApiRequest request = ApiRequestFactory.GetApiRequest(AccessToken);
-            string endpoint = Endpoints.TextTrack;
+            var request = _apiRequestFactory.GetApiRequest(AccessToken);
+            const string endpoint = Endpoints.TextTrack;
             request.Method = HttpMethod.Delete;
             request.Path = endpoint;
 
